requirements.txt
pandas
streamlit
python-dotenv
genai
streamlit_option_menu 
streamlit_lottie
streamlit_chat
africastalking
openai
seaborn
pyttsx3
gtts
requests